+++ python/qx_utilities/processing/simple.py
@@ -18,7 +18,7 @@ It consists of functions:
 All the functions are part of the processing suite. They should be called from
 the command line using `qunex` command. Help is available through:
 
-- `qunex ?<command>` for command specific help
+- `qunex <command> --help` for command specific help
 """
 
 # Created by Grega Repovs on 2016-12-17.